.devcontainer/requirements.txt
python-dotenv
numpy>=1.26.4
pandas==2.1.2
matplotlib>=3.8
plotly==5.18.0
yfinance==0.2.31
cufflinks==0.17.3
jupyter==1.0.0
jupyterlab==4.0.7
notebook==7.0.6
ipython==8.16.1
pandas-stubs
blosc2>=2.3.0
pytz==2023.3.post1
requests==2.31.0
python-dateutil==2.8.2
arrow==1.3.0
beautifulsoup4==4.12.2
lxml==4.9.3
PyYAML==6.0.1
python-json-logger==2.0.7
debugpy==1.8.0
prompt-toolkit==3.0.39
tqdm==4.66.1
click==8.1.7
pyyaml